Feat: implement login endpoint with lockout and token issuance
Add POST /auth/login: validates password, returns a generic 401 for
unknown email, wrong password, or a locked account (no way to tell
them apart from the response), and locks the account for 15 minutes
after 5 failed attempts. Add UnauthorizedException/ForbiddenException
mapped to 401/403, and issueTokens to sign the JWT access token and
persist the first refresh token row on successful login.

Fix: quarkus-flyway-postgresql resolved as a plain jar but was never
registered as a real Quarkus extension, so migrate-at-start was
silently ignored and no migration ever ran. Replaced with
quarkus-flyway + flyway-database-postgresql per the official setup.
Also fixes two column type mismatches (token_hash, failed_login_attempts)
caught once Flyway started running for real, and removes the leftover
MyEntity archetype boilerplate that was blocking schema validation.

+++ src/main/resources/db/migration/V1__create_system_users.sql
@@ -4,7 +4,7 @@ CREATE TABLE system_users (
                               password_hash VARCHAR(255) NOT NULL,
                               status VARCHAR(20) NOT NULL DEFAULT 'ACTIVE'
                                   CHECK (status IN ('ACTIVE', 'LOCKED', 'DISABLED')),
-                              failed_login_attempts SMALLINT NOT NULL DEFAULT 0,
+                              failed_login_attempts INTEGER NOT NULL DEFAULT 0,
                               locked_until TIMESTAMP,
                               last_login_at TIMESTAMP,
                               password_changed_at TIMESTAMP,
